Update README.md
I believe it is important not to miss out GDPR, as the most comprehensive, relevant and up-to-date legal source.
Maybe the regulation itself in some cases seems to be a bit too "legal", but there are some really useful, easy to understand and reliable sources. Eg. one of the bests is ICO's guide: https://ico.org.uk/for-organisations/guide-to-data-protection/